Add handling for `dct:applicableLegislation`
While this is required by the BReg spec, as the `dct:type` field is made
mandatory on `eli:LegalResource` therein, no field that we're handling as
part of BReg requires an `eli:LegalResource`, so I haven't put it in the
BReg profile.

The list of included ELIs will likely need to change.